# server/genapi/model_receive_response.go
/*
iPubSub API

A lightweight, scalable pub-sub service API that supports both real-time message delivery and persistent storage.  iPubSub enables publishers to send messages to lightweight topics/streams, and subscribers to receive messages  using efficient long-polling. Topics are created dynamically on first message without explicit provisioning.  **Core Features:** - **Real-time Matching**: Publishers and subscribers are matched in real-time using long polling - **Dynamic Topics**: Lightweight topics/streams created automatically on first message - **Dual Storage**: Messages can be delivered in-memory for real-time consumption and/or persisted for replay - **Per-message TTL**: Individual message expiration (not stream-level) - **Horizontal Scaling**: Distributed hash ring for stream routing across nodes

API version: 1.0.0
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package genapi

import (
	"bytes"
	"encoding/json"
	"fmt"
	"time"
)

// checks if the ReceiveResponse type satisfies the MappedNullable interface at compile time
var _ MappedNullable = &ReceiveResponse{}

// ReceiveResponse struct for ReceiveResponse
type ReceiveResponse struct {
	// Unique identifier for the message
	MessageUuid *string `json:"messageUuid,omitempty"`
	// The received message data (can be any JSON value - object, array, string, number, boolean, or null)
	Message interface{} `json:"message"`
	// When the output was generated
	Timestamp time.Time `json:"timestamp"`
	// Token for resuming from next position, only applicable when writeToDB is true.
	DbResumeToken *string `json:"dbResumeToken,omitempty"`
}

type _ReceiveResponse ReceiveResponse

// NewReceiveResponse instantiates a new ReceiveResponse object
// This constructor will assign default values to properties that have it defined,
// and makes sure properties required by API are set, but the set of arguments
// will change when the set of required properties is changed
func NewReceiveResponse(message interface{}, timestamp time.Time) *ReceiveResponse {
	this := ReceiveResponse{}
	this.Message = message
	this.Timestamp = timestamp
	return &this
}

// NewReceiveResponseWithDefaults instantiates a new ReceiveResponse object
// This constructor will only assign default values to properties that have it defined,
// but it doesn't guarantee that properties required by API are set
func NewReceiveResponseWithDefaults() *ReceiveResponse {
	this := ReceiveResponse{}
	return &this
}

// GetMessageUuid returns the MessageUuid field value if set, zero value otherwise.
func (o *ReceiveResponse) GetMessageUuid() string {
	if o == nil || IsNil(o.MessageUuid) {
		var ret string
		return ret
	}
	return *o.MessageUuid
}

// GetMessageUuidOk returns a tuple with the MessageUuid field value if set, nil otherwise
// and a boolean to check if the value has been set.
func (o *ReceiveResponse) GetMessageUuidOk() (*string, bool) {
	if o == nil || IsNil(o.MessageUuid) {
		return nil, false
	}
	return o.MessageUuid, true
}

// HasMessageUuid returns a boolean if a field has been set.
func (o *ReceiveResponse) HasMessageUuid() bool {
	if o != nil && !IsNil(o.MessageUuid) {
		return true
	}

	return false
}

// SetMessageUuid gets a reference to the given string and assigns it to the MessageUuid field.
func (o *ReceiveResponse) SetMessageUuid(v string) {
	o.MessageUuid = &v
}

// GetMessage returns the Message field value
// If the value is explicit nil, the zero value for interface{} will be returned
func (o *ReceiveResponse) GetMessage() interface{} {
	if o == nil {
		var ret interface{}
		return ret
	}

	return o.Message
}

// GetMessageOk returns a tuple with the Message field value
// and a boolean to check if the value has been set.
// NOTE: If the value is an explicit nil, `nil, true` will be returned
func (o *ReceiveResponse) GetMessageOk() (*interface{}, bool) {
	if o == nil || IsNil(o.Message) {
		return nil, false
	}
	return &o.Message, true
}

// SetMessage sets field value
func (o *ReceiveResponse) SetMessage(v interface{}) {
	o.Message = v
}

// GetTimestamp returns the Timestamp field value
func (o *ReceiveResponse) GetTimestamp() time.Time {
	if o == nil {
		var ret time.Time
		return ret
	}

	return o.Timestamp
}

// GetTimestampOk returns a tuple with the Timestamp field value
// and a boolean to check if the value has been set.
func (o *ReceiveResponse) GetTimestampOk() (*time.Time, bool) {
	if o == nil {
		return nil, false
	}
	return &o.Timestamp, true
}

// SetTimestamp sets field value
func (o *ReceiveResponse) SetTimestamp(v time.Time) {
	o.Timestamp = v
}

// GetDbResumeToken returns the DbResumeToken field value if set, zero value otherwise.
func (o *ReceiveResponse) GetDbResumeToken() string {
	if o == nil || IsNil(o.DbResumeToken) {
		var ret string
		return ret
	}
	return *o.DbResumeToken
}

// GetDbResumeTokenOk returns a tuple with the DbResumeToken field value if set, nil otherwise
// and a boolean to check if the value has been set.
func (o *ReceiveResponse) GetDbResumeTokenOk() (*string, bool) {
	if o == nil || IsNil(o.DbResumeToken) {
		return nil, false
	}
	return o.DbResumeToken, true
}

// HasDbResumeToken returns a boolean if a field has been set.
func (o *ReceiveResponse) HasDbResumeToken() bool {
	if o != nil && !IsNil(o.DbResumeToken) {
		return true
	}

	return false
}

// SetDbResumeToken gets a reference to the given string and assigns it to the DbResumeToken field.
func (o *ReceiveResponse) SetDbResumeToken(v string) {
	o.DbResumeToken = &v
}

func (o ReceiveResponse) MarshalJSON() ([]byte, error) {
	toSerialize, err := o.ToMap()
	if err != nil {
		return []byte{}, err
	}
	return json.Marshal(toSerialize)
}

func (o ReceiveResponse) ToMap() (map[string]interface{}, error) {
	toSerialize := map[string]interface{}{}
	if !IsNil(o.MessageUuid) {
		toSerialize["messageUuid"] = o.MessageUuid
	}
	if o.Message != nil {
		toSerialize["message"] = o.Message
	}
	toSerialize["timestamp"] = o.Timestamp
	if !IsNil(o.DbResumeToken) {
		toSerialize["dbResumeToken"] = o.DbResumeToken
	}
	return toSerialize, nil
}

func (o *ReceiveResponse) UnmarshalJSON(data []byte) (err error) {
	// This validates that all required properties are included in the JSON object
	// by unmarshalling the object into a generic map with string keys and checking
	// that every required field exists as a key in the generic map.
	requiredProperties := []string{
		"message",
		"timestamp",
	}

	allProperties := make(map[string]interface{})

	err = json.Unmarshal(data, &allProperties)

	if err != nil {
		return err
	}

	for _, requiredProperty := range requiredProperties {
		if _, exists := allProperties[requiredProperty]; !exists {
			return fmt.Errorf("no value given for required property %v", requiredProperty)
		}
	}

	varReceiveResponse := _ReceiveResponse{}

	decoder := json.NewDecoder(bytes.NewReader(data))
	decoder.DisallowUnknownFields()
	err = decoder.Decode(&varReceiveResponse)

	if err != nil {
		return err
	}

	*o = ReceiveResponse(varReceiveResponse)

	return err
}

type NullableReceiveResponse struct {
	value *ReceiveResponse
	isSet bool
}

func (v NullableReceiveResponse) Get() *ReceiveResponse {
	return v.value
}

func (v *NullableReceiveResponse) Set(val *ReceiveResponse) {
	v.value = val
	v.isSet = true
}

func (v NullableReceiveResponse) IsSet() bool {
	return v.isSet
}

func (v *NullableReceiveResponse) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableReceiveResponse(val *ReceiveResponse) *NullableReceiveResponse {
	return &NullableReceiveResponse{value: val, isSet: true}
}

func (v NullableReceiveResponse) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableReceiveResponse) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}